+++ src/app/views/PersonInfo.svelte
@@ -1,7 +1,12 @@
 <script lang="ts">
   import * as nip19 from "nostr-tools/nip19"
   import {Router} from "@welshman/router"
+  import {parseJson} from "@welshman/lib"
   import {deriveProfile, deriveHandleForPubkey, displayHandle} from "@welshman/app"
+  import {copyToClipboard} from "src/util/html"
+  import {showInfo} from "src/partials/Toast.svelte"
+  import Field from "src/partials/Field.svelte"
+  import Anchor from "src/partials/Anchor.svelte"
   import CopyValue from "src/partials/CopyValue.svelte"
   import RelayCard from "src/app/shared/RelayCard.svelte"
 
@@ -11,6 +16,12 @@
   const handle = deriveHandleForPubkey(pubkey)
   const relays = Router.get().FromPubkey(pubkey).getUrls()
 
+  const copyJson = () => {
+    copyToClipboard(json)
+    showInfo(`Profile JSON copied to clipboard!`)
+  }
+
+  $: json = JSON.stringify(parseJson($profile?.event?.content), null, 2)
   $: lightningAddress = $profile?.lud16 || $profile?.lud06
 </script>
 
@@ -42,4 +53,15 @@
     <i class="fa-solid fa-info-circle" />
     No lightning address found.
   </p>
+{/if}
+{#if $profile}
+  <Field>
+    <p slot="label">Profile JSON</p>
+    <div class="relative rounded bg-tinted-700 p-1">
+      <pre class="overflow-auto text-xs"><code>{json}</code></pre>
+      <Anchor circle class="absolute right-1 top-1 bg-neutral-800" on:click={copyJson}>
+        <i class="fa fa-copy m-2" />
+      </Anchor>
+    </div>
+  </Field>
 {/if}